backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
paho-mqtt==1.6.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
